requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
pypdf
faiss-cpu
langchain_google_genai